# apps/api/openapi.yml
openapi: 3.0.0
paths:
...
  /users:
    post:
      operationId: UsersController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResponsePrimary"
      tags: &a1
        - users
    patch:
      operationId: UsersController_updateMe
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserInfo"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResponsePrimary"
      tags: *a1
      security:
        - bearer: []
    delete:
      operationId: UsersController_deleteMe
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MiniUserResponse"
      tags: *a1
      security:
        - bearer: []
    get:
      operationId: UsersController_findAll
      parameters:
        - name: field
          required: false
          in: query
          schema:
            type: string
        - name: order
          required: false
          in: query
          schema:
            type: string
        - name: returnFullCount
          required: false
          in: query
          schema:
            type: boolean
        - name: responseType
          required: false
          in: query
          schema:
            type: string
        - name: page
          required: false
          in: query
          schema:
            type: number
        - name: pageSize
          required: false
          in: query
          schema:
            type: number
        - name: name
          required: false
          in: query
          schema:
            type: string
        - name: username
          required: false
          in: query
          schema:
            type: string
        - name: email
          required: false
          in: query
          schema:
            type: string
        - name: country
          required: false
          in: query
          schema:
            type: string
        - name: location
          required: false
          in: query
          schema:
            type: string
        - name: age
          required: false
          in: query
          schema:
            type: number
      responses:
        "200":
          content:
            application/json:
              examples:
                UserResponse:
                  value:
                    metadata:
                      pagination:
                        page: 1
                        pageSize: 10
                      links:
                        first: https://example.com/api/users?page=1
                        prev: https://example.com/api/users?page=0
                        next: https://example.com/api/users?page=2
                      query:
                        page: 1
                        pageSize: 10
                        username: john_doe
                        country: USA
                        field: username
                        order: asc
                        responseType: mini
                    results:
                      - id: 1
                        username: john_doe
                        isFake: false
                        profilePicture: https://example.com/image.jpg
                        country: USA
                        bio: I am a cool person
                        email: john.doe@example.com
                        followers: 0
                        level: 0
                        name: John Doe
                        age: 20
                        updatedAt: &a2 2025-02-26T13:38:22.138Z
                MiniUserResponse:
                  value:
                    metadata:
                      pagination:
                        page: 1
                        pageSize: 10
                      links:
                        first: https://example.com/api/users?page=1
                        prev: https://example.com/api/users?page=0
                        next: https://example.com/api/users?page=2
                      query:
                        page: 1
                        pageSize: 10
                        username: john_doe
                        country: USA
                        field: username
                        order: asc
                        responseType: mini
                    results:
                      - id: 1
                        username: john_doe
                MiniUserResponseWithCountry:
                  value:
                    metadata:
                      pagination:
                        page: 1
                        pageSize: 10
                      links:
                        first: https://example.com/api/users?page=1
                        prev: https://example.com/api/users?page=0
                        next: https://example.com/api/users?page=2
                      query:
                        page: 1
                        pageSize: 10
                        username: john_doe
                        country: USA
                        field: username
                        order: asc
                        responseType: mini
                    results:
                      - id: 1
                        username: john_doe
                        isFake: false
                        profilePicture: https://example.com/image.jpg
                        country: USA
                MiniUserResponseWithProfilePicture:
                  value:
                    metadata:
                      pagination:
                        page: 1
                        pageSize: 10
                      links:
                        first: https://example.com/api/users?page=1
                        prev: https://example.com/api/users?page=0
                        next: https://example.com/api/users?page=2
                      query:
                        page: 1
                        pageSize: 10
                        username: john_doe
                        country: USA
                        field: username
                        order: asc
                        responseType: mini
                    results:
                      - id: 1
                        username: john_doe
                        isFake: false
                        profilePicture: https://example.com/image.jpg
                ExtendedUserResponse:
                  value:
                    metadata:
                      pagination:
                        page: 1
                        pageSize: 10
                      links:
                        first: https://example.com/api/users?page=1
                        prev: https://example.com/api/users?page=0
                        next: https://example.com/api/users?page=2
                      query:
                        page: 1
                        pageSize: 10
                        username: john_doe
                        country: USA
                        field: username
                        order: asc
                        responseType: mini
                    results:
                      - id: 1
                        username: john_doe
                        isFake: false
                        profilePicture: https://example.com/image.jpg
                        country: USA
                        bio: I am a cool person
                        email: john.doe@example.com
                        followers: 0
                        level: 0
                        name: John Doe
                        age: 20
                        updatedAt: *a2
                        createdAt: 2025-02-26T13:38:22.138Z
                        following: 0
                        location: New York
          description: ""
      tags: *a1
  /users/me:
    get:
      operationId: UsersController_findMe
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtendedUserResponse"
      tags: *a1
      security:
        - bearer: []
  /users/{id}:
    get:
      operationId: UsersController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
        - name: responseType
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: "#/components/schemas/MiniUserResponse"
                  - $ref: "#/components/schemas/UserResponse"
                  - $ref: "#/components/schemas/ExtendedUserResponse"
                  - $ref: "#/components/schemas/MiniUserResponseWithCountry"
                  - $ref: "#/components/schemas/MiniUserResponseWithProfilePicture"
              examples:
                UserResponse:
                  value:
                    id: 1
                    username: john_doe
                    isFake: false
                    profilePicture: https://example.com/image.jpg
                    country: USA
                    bio: I am a cool person
                    email: john.doe@example.com
                    followers: 0
                    level: 0
                    name: John Doe
                    age: 20
                    updatedAt: &a3 2025-02-26T13:38:22.138Z
                MiniUserResponse:
                  value:
                    id: 1
                    username: john_doe
                MiniUserResponseWithCountry:
                  value:
                    id: 1
                    username: john_doe
                    isFake: false
                    profilePicture: https://example.com/image.jpg
                    country: USA
                MiniUserResponseWithProfilePicture:
                  value:
                    id: 1
                    username: john_doe
                    isFake: false
                    profilePicture: https://example.com/image.jpg
                ExtendedUserResponse:
                  value:
                    id: 1
                    username: john_doe
                    isFake: false
                    profilePicture: https://example.com/image.jpg
                    country: USA
                    bio: I am a cool person
                    email: john.doe@example.com
                    followers: 0
                    level: 0
                    name: John Doe
                    age: 20
                    updatedAt: *a3
                    createdAt: 2025-02-26T13:38:22.138Z
                    following: 0
                    location: New York
          description: ""
      tags: *a1
    patch:
      operationId: UsersController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserInfo"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResponsePrimary"
      tags: *a1
      security:
        - bearer: []
    delete:
      operationId: UsersController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResponsePrimary"
      tags: *a1
      security:
        - bearer: []
  /users/fake:
    post:
      operationId: UsersController_createFake
      parameters: []
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResponsePrimary"
      tags: *a1
      security:
        - bearer: []
...
info:
  title: winning.sh API
  description: The Winning.sh API description
  version: 1.0 alpha
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    MiniUserResponse:
      type: object
      properties:
        id:
          type: number
          readOnly: true
        username:
          type: string
          readOnly: true
        isFake:
          type: boolean
          readOnly: true
      required:
        - id
        - username
        - isFake
    MiniUserResponseWithProfilePicture:
      type: object
      properties:
        id:
          type: number
          readOnly: true
        username:
          type: string
          readOnly: true
        isFake:
          type: boolean
          readOnly: true
        profilePicture:
          type: string
          readOnly: true
      required:
        - id
        - username
        - isFake
        - profilePicture
    MiniUserResponseWithCountry:
      type: object
      properties:
        id:
          type: number
          readOnly: true
        username:
          type: string
          readOnly: true
        isFake:
          type: boolean
          readOnly: true
        profilePicture:
          type: string
          readOnly: true
        country:
          type: string
          readOnly: true
      required:
        - id
        - username
        - isFake
        - profilePicture
        - country
    UserResponse:
      type: object
      properties:
        id:
          type: number
          readOnly: true
        username:
          type: string
          readOnly: true
        isFake:
          type: boolean
          readOnly: true
        profilePicture:
          type: string
          readOnly: true
        country:
          type: string
          readOnly: true
        email:
          type: string
          readOnly: true
        bio:
          type: string
          readOnly: true
        level:
          type: number
          readOnly: true
        name:
          type: string
          readOnly: true
        age:
          type: number
          readOnly: true
        updatedAt:
          format: date-time
          type: string
          readOnly: true
        followers:
          type: number
          readOnly: true
      required:
        - id
        - username
        - isFake
        - profilePicture
        - country
        - email
        - bio
        - level
        - name
        - age
        - updatedAt
        - followers
    ExtendedUserResponse:
      type: object
      properties:
        id:
          type: number
          readOnly: true
        username:
          type: string
          readOnly: true
        isFake:
          type: boolean
          readOnly: true
        profilePicture:
          type: string
          readOnly: true
        country:
          type: string
          readOnly: true
        email:
          type: string
          readOnly: true
        bio:
          type: string
          readOnly: true
        level:
          type: number
          readOnly: true
        name:
          type: string
          readOnly: true
        age:
          type: number
          readOnly: true
        updatedAt:
          format: date-time
          type: string
          readOnly: true
        followers:
          type: number
          readOnly: true
        location:
          type: string
          readOnly: true
        following:
          type: number
          readOnly: true
        createdAt:
          format: date-time
          type: string
          readOnly: true
      required:
        - id
        - username
        - isFake
        - profilePicture
        - country
        - email
        - bio
        - level
        - name
        - age
        - updatedAt
        - followers
        - location
        - following
        - createdAt
...
    CreateUserRequest:
      type: object
      properties:
        name:
          type: string
        username:
          type: string
        bio:
          type: string
        email:
          type: string
        password:
          type: string
        profilePicture:
          type: string
        country:
          type: string
        dateOfBirth:
          format: date-time
          type: string
      required:
        - name
        - username
        - email
        - password
        - country
        - dateOfBirth
    ActionResponsePrimary:
      type: object
      properties:
        id:
          type: number
          readOnly: true
      required:
        - id
    UpdateUserInfo:
      type: object
      properties:
        profilePicture:
          type: string
        country:
          type: string
        username:
          type: string
        name:
          type: string
        bio:
          type: string
        dateOfBirth:
          format: date-time
          type: string
        location:
          type: string
...